[Mailer] Move abstract test case to Tests/
This PR

is based on https://github.com/symfony/symfony/pull/39495/files#r544203249

cc @wouterj @nicolas-grekas
