+++ .github/composer-config.json
@@ -4,6 +4,7 @@
         "preferred-install": {
             "symfony/form": "source",
             "symfony/http-kernel": "source",
+            "symfony/mailer": "source",
             "symfony/validator": "source",
             "*": "dist"
         }
