+++ src/Symfony/Component/Mailer/Bridge/Sendgrid/composer.json
@@ -20,7 +20,7 @@
         "symfony/mailer": "^4.4|^5.0"
     },
     "require-dev": {
-        "symfony/http-client": "^4.3|^5.0"
+        "symfony/mailer": "^4.4.18|^5.1.10"
     },
     "autoload": {
         "psr-4": { "Symfony\\Component\\Mailer\\Bridge\\Sendgrid\\": "" },
