+++ src/Symfony/Component/Mailer/Test/Constraint/EmailCount.php
@@ -9,7 +9,7 @@
  * file that was distributed with this source code.
  */
 
-namespace Symfony\Component\Mailer\Test\Constraint;
+namespace Symfony\Component\Mailer\Tests\Constraint;
 
 use PHPUnit\Framework\Constraint\Constraint;
 use Symfony\Component\Mailer\Event\MessageEvents;
