+++ src/Symfony/Component/Mailer/Tests/Transport/SendmailTransportFactoryTest.php
@@ -11,7 +11,7 @@
 
 namespace Symfony\Component\Mailer\Tests\Transport;
 
-use Symfony\Component\Mailer\Test\TransportFactoryTestCase;
+use Symfony\Component\Mailer\Tests\TransportFactoryTestCase;
 use Symfony\Component\Mailer\Transport\Dsn;
 use Symfony\Component\Mailer\Transport\SendmailTransport;
 use Symfony\Component\Mailer\Transport\SendmailTransportFactory;
